Circular imports gone, merged db_utils into utils.
All utils are now happily in one file.
