+++ amivapi/utils.py
@@ -10,7 +10,21 @@ from os import urandom
 import smtplib
 from email.mime.text import MIMEText
 from copy import deepcopy
+import datetime as dt
+import json
 
+from sqlalchemy.types import TypeDecorator
+from sqlalchemy.exc import InvalidRequestError
+from sqlalchemy.ext.declarative import declarative_base, declared_attr
+from sqlalchemy.orm import synonym
+from sqlalchemy import (
+    Column,
+    Text,
+    String,
+    Integer,
+    ForeignKey,
+    DateTime,
+)
 from flask import Config, g, current_app as app
 
 from eve.utils import config, request_method
@@ -129,6 +143,73 @@ def mail(sender, to, subject, text):
 EMAIL_REGEX = '^.+@.+$'
 
 
+class ValidatorAMIV(ValidatorSQL):
+    """Validator subclass adding more validation for special fields."""
+
+    def _validate_not_patchable(self, enabled, field, value):
+        """Custom Validator to inhibit patching of the field.
+
+        e.g. eventsignups, userid: required for post, but can not be patched
+
+        Args:
+            enabled (bool): Boolean, should be true
+            field (string): field name.
+            value: field value.
+        """
+        if enabled and (request_method() == 'PATCH'):
+            self._error(field, "this field can not be changed with PATCH")
+
+    def _validate_not_patchable_unless_admin(self, enabled, field, value):
+        """Inhibit patching of the field.
+
+        e.g. eventsignups, userid: required for post, but can not be patched
+
+        Args:
+            enabled (bool): Boolean, should be true
+            field (string): field name.
+            value: field value.
+        """
+        if enabled and (request_method() == 'PATCH') and not g.resource_admin:
+            self._error(field, "this field can not be changed with PATCH "
+                        "unless you have admin rights.")
+
+    def _validate_unique_combination(self, unique_combination, field, value):
+        """Validate that a combination of fields is unique.
+
+        e.g. user with id 1 can have several eventsignups for different events,
+        but only 1 eventsignup for event with id 42
+
+        unique_combination should be a list of other fields
+
+        Note: Make sure that other fields actually exists (setting them to
+        required etc)
+
+        Args:
+            unique_combination (list): combination fields
+            field (string): field name.
+            value: field value.
+        """
+        lookup = {field: value}  # self
+        for other_field in unique_combination:
+            lookup[other_field] = self.document.get(other_field)
+
+        # If we are patching the issue is more complicated, some fields might
+        # have to be checked but are not part of the document because they will
+        # not be patched. We have to load them from the database
+        patch = (request_method() == 'PATCH')
+        if patch:
+            original = self._original_document
+            for key in unique_combination:
+                if key not in self.document.keys():
+                    lookup[key] = original[key]
+
+        # Now check database
+        if app.data.find_one(self.resource, None, **lookup) is not None:
+            self._error(field, "value already exists in the database in " +
+                        "combination with values for: %s" %
+                        unique_combination)
+
+
 def make_domain(model):
     """Make Eve domain for a model.
 
@@ -208,70 +289,152 @@ def register_validator(app, validator_class):
     app.validator = type("Adopted_%s" % validator_class.__name__,
                          (validator_class, app.validator),
                          {})
+# db utils
 
 
-class ValidatorAMIV(ValidatorSQL):
-    """Validator subclass adding more validation for special fields."""
+class JSON(TypeDecorator):
+    """Column type to transparently handle JSONified content.
 
-    def _validate_not_patchable(self, enabled, field, value):
-        """Custom Validator to inhibit patching of the field.
+    Converts between python objects and a String column in the database.
 
-        e.g. eventsignups, userid: required for post, but can not be patched
+    Usage:
+        JSON(255) (implies a String(255) column)
 
-        Args:
-            enabled (bool): Boolean, should be true
-            field (string): field name.
-            value: field value.
+    See also:
+        http://docs.sqlalchemy.org/en/rel_0_7/core/types.html#marshal-json-strings
+    """
+
+    impl = String
+
+    def process_bind_param(self, value, dialect):
+        """Application -> Database."""
+        if value is not None:
+            value = json.dumps(value)
+
+        return value
+
+    def process_result_value(self, value, dialect):
+        """Database -> Application."""
+        if value is not None:
+            try:
+                value = json.loads(value)
+            except (TypeError, ValueError):
+                raise ValueError("Invalid JSON found in database: %r", value)
+
+        return value
+
+
+class JSONText(JSON):
+    """Column to handle variable length JSON."""
+
+    impl = Text
+
+
+class BaseModel(object):
+    """Mixin for common columns."""
+
+    # __abstract__ makes SQL Alchemy not create a table for this class
+    __abstract__ = True
+
+    # All classes which overwrite this with True will get exposed as a resource
+    __expose__ = False
+
+    # For documentation
+    __description__ = {}
+
+    # This is a list of fields, which are added to the projection created by
+    # eve. Add any additional field to be delivered on GET requests by default
+    # in the subclasses.
+    __projected_fields__ = []
+    # These fields are embedded by default meaning that not only the related
+    # ids are used, but the whole object included
+    __embedded_fields__ = []
+
+    # These can contain a list of methods which need some kind of
+    # authorization. If nothing is set only admins can access the method
+    __public_methods__ = []
+    __owner_methods__ = []
+    __registered_methods__ = []
+    __owner__ = []
+
+    @declared_attr
+    def __tablename__(cls):
+        """Make name for table."""
+        # Correct English attaches 'es' to plural forms which end in 's'
+        if cls.__name__.lower()[-1:] == 's':
+            return "%ses" % cls.__name__.lower()
+        return "%ss" % cls.__name__.lower()
+
+    id = Column(Integer, primary_key=True, autoincrement=True)
+
+    @declared_attr
+    def _id(cls):
+        return synonym("id")
+
+    _created = Column(DateTime, default=dt.datetime.utcnow)
+    _updated = Column(
+        DateTime, default=dt.datetime.utcnow, onupdate=dt.datetime.utcnow)
+    _etag = Column(String(50))
+
+    # This needs to be a function, as it has a ForeignKey in a mixin. The
+    # function makes binding at a later point possible
+    @declared_attr
+    def _author(cls):
+        return Column(Integer, ForeignKey("users.id"))  # , nullable=False)
+
+    @declared_attr
+    def __self__(cls):
+        r"""Make own class available to models.
+
+        This makes it possible to reference fields indirectly with
+        User.__self__.sessions. We use this as a workaround, as the lookup dict
+        for eve can not contain sqlalchemy operators for a relation.
+
+            { "sessions": "indirect_any(\"token\")" }
+
+        This would fail cause eve_sqlalchemy first checks wether sessions is a
+        relation and therefore never reaches the custom operator clause.
+
+            { "__self__": "indirect_any(\"sessions.token\")" }
+
+        This works as __self__ is not a relation.
         """
-        if enabled and (request_method() == 'PATCH'):
-            self._error(field, "this field can not be changed with PATCH")
+        return cls
 
-    def _validate_not_patchable_unless_admin(self, enabled, field, value):
-        """Inhibit patching of the field.
+    @classmethod
+    def indirect_any(cls, querystring):
+        r"""Check if any field is matched over various relations.
 
-        e.g. eventsignups, userid: required for post, but can not be patched
+        This is used to check indirect permissions in authorization.py
 
-        Args:
-            enabled (bool): Boolean, should be true
-            field (string): field name.
-            value: field value.
+        When the parser hits one to many relations it will create any clauses
+
+        Example:
+            User.recursive_any("groups.addresses.address, \"it@amiv.ethz.ch\"")
+
+            this will return all users, which are in a group which receives
+            emails from it@amiv.ethz.ch
+
+        @argument querystring: String of two comma seperated arguments. The
+                               first is the field which is to be checked. The
+                               second is the value to match.
+
+        @returns: sqlalchemy statement
         """
-        if enabled and (request_method() == 'PATCH') and not g.resource_admin:
-            self._error(field, "this field can not be changed with PATCH "
-                        "unless you have admin rights.")
+        field, value = map(unicode.strip, querystring.split(','))
 
-    def _validate_unique_combination(self, unique_combination, field, value):
-        """Validate that a combination of fields is unique.
+        parts = field.strip().split('.')
+        relation = getattr(cls, parts[0])
 
-        e.g. user with id 1 can have several eventsignups for different events,
-        but only 1 eventsignup for event with id 42
+        try:
+            # This will work for many-to-one relationships
+            return relation.has(**{'.'.join(parts[1:]): value})
+        except InvalidRequestError:
+            # This will work for one-to-many relationships
+            # In this case we accept if there is any related object which
+            # satisfies the requirement
+            return relation.any(**{'.'.join(parts[1:]): value})
 
-        unique_combination should be a list of other fields
 
-        Note: Make sure that other fields actually exists (setting them to
-        required etc)
-
-        Args:
-            unique_combination (list): combination fields
-            field (string): field name.
-            value: field value.
-        """
-        lookup = {field: value}  # self
-        for other_field in unique_combination:
-            lookup[other_field] = self.document.get(other_field)
-
-        # If we are patching the issue is more complicated, some fields might
-        # have to be checked but are not part of the document because they will
-        # not be patched. We have to load them from the database
-        patch = (request_method() == 'PATCH')
-        if patch:
-            original = self._original_document
-            for key in unique_combination:
-                if key not in self.document.keys():
-                    lookup[key] = original[key]
-
-        # Now check database
-        if app.data.find_one(self.resource, None, **lookup) is not None:
-            self._error(field, "value already exists in the database in " +
-                        "combination with values for: %s" %
-                        unique_combination)
+Base = declarative_base(cls=BaseModel)
+"""Declarative base to use for all models."""
